# artifacts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/messages/threads/{id}/attachments": {
      "post": {
        "operationId": "MessagesController_uploadAttachment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "LankaDWS API",
    "description": "Permission-scoped LankaDWS modular monolith API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lankadws_session"
      }
    },
...
  }
}